+++ models/marts/grades/fct_student_section_grade.sql
@@ -25,5 +25,14 @@ SELECT
     grading_period_reference.school_year                                    AS school_year,
     numeric_grade_earned                                                    AS numeric_grade_earned,
     letter_grade_earned                                                     AS letter_grade_earned,
-    grade_type_descriptor                                                   AS grade_type
+    grade_type_descriptor                                                   AS grade_type,
+    IF(CURRENT_DATE BETWEEN ssa.begin_date AND ssa.end_date, 1, 0)          AS is_actively_enrolled_in_section
 FROM {{ ref('stg_edfi_grades') }} grades
+LEFT JOIN {{ ref('stg_edfi_student_section_associations') }} ssa
+    ON grades.school_year = ssa.school_year
+    AND grades.student_section_association_reference.student_unique_id = ssa.student_reference.student_unique_id
+    AND grades.student_section_association_reference.begin_date = ssa.begin_date
+    AND grades.student_section_association_reference.local_course_code = ssa.section_reference.local_course_code
+    AND grades.student_section_association_reference.school_id = ssa.section_reference.school_id
+    AND grades.student_section_association_reference.school_year = ssa.section_reference.school_year
+    AND grades.student_section_association_reference.session_name = ssa.section_reference.session_name